+++ pkgs/fc/agent/fc/manage/tests/test_postgresql.py
@@ -1,5 +1,6 @@
 import json
 import traceback
+from typing import List
 from unittest.mock import Mock, patch
 
 import fc.manage.postgresql